Delete 'draft' return versions https://eaflood.atlassian.net/browse/WATER-5160 As part of taking over returns management from NALD, we needed to import all the return versions and requirements data.
However, what we hadn't spotted was that NALD has the principle of ‘draft’ return versions. The reporting work stream has spotted 10 errant return versions, which, after investigation, are ‘draft’ NALD return versions.

So in this PR we will create a data fix migration to delete the 10 draft return versions and any associated child records. Whilst ensuring that the return version history for the licence is left valid.