Remove XML upload feature and dependencies https://eaflood.atlassian.net/browse/WATER-5138
Uninstall the `libxmljs2` dependency and remove any modules that use it, reducing the feature as much as necessary.

The functionality and all traces of it do not need to be removed completely. Our approach to the legacy apps is to do as little as possible, in the knowledge they will be made redundant. It also reduces the risk of deleting something we later find another part of the code is dependent on.