Switch to using Snyk CLI action in CI
https://github.com/DEFRA/water-abstraction-team/issues/144

With the adoption of SSO in the Defra GitHub org our current [Snyk](https://snyk.io) status checks have stopped working.

The team issue goes into more detail, but in essence whilst the GitHub account that **Snyk** uses to connect has SSO enabled, it won't connect to GitHub to update the PR with the test result.

This is a 'test' of trying to perform the same check without **Snyk** having to authenticate to GitHub. Instead, GitHub will authenticate with **Snyk** by using its CLI via a GitHub action.

It does mean **Snyk** will no longer be listed as a separate check, instead it will just be a new step in our CI. So, the outcome (adding a package with a vulnerability will cause the PR to fail checks) will be the same.