Add .snyk config file
Turns out the Snyk CLI does not have the ability to test on the change in the PR. Instead, it will test the whole repo, which leads to it failing the CI.

We know we have issues with the project in its current state. This check is about ensuring we don't make it any worse. So, the only solution is to add a config file to tell the CLI to ignore the issues we know already exist.

For reference, this was generated by first install the [Snyk CLI](https://docs.snyk.io/developer-tools/snyk-cli/install-or-update-the-snyk-cli), then running `snyk test` locally.

We then extracted the Snyk issue IDs and used the [snyk ignore](https://docs.snyk.io/developer-tools/snyk-cli/commands/ignore) command to create the file and entries (first command created the file. Subsequent commands added new entries to it). For example, `snyk ignore --id=SNYK-JS-LIBXMLJS2-6808810 --reason='No direct upgrade or patch'`.