Fix background jobs race condition (#2708)
https://eaflood.atlassian.net/browse/WATER-5174

Users reported that customer file name and date weren't being displayed in the UI. We found that this was because the underlying data wasn't being refreshed in the db. However, there was no obvious reason _why_ it wasn't being refreshed!

Checking the logs revealed that the job responsible for this (`billing.customer-file-refresh`) was being registered and added to `QueueManager` by `background-service`, but never actually ran. This turned out to be because:

- `background-service` and `water-api` both register the job to the queue;
- As part of this process, they clear any existing `billing.customer-file-refresh` jobs from the queue because the job has the `startClean` flag set (which we use to ensure we only have one instance of the job scheduled to run hourly);
- Only `background-service` adds the job to the queue;
- And therefore, if `water-api` registers the job _after_ `background-service`, the existing legitimate `billing.customer-file-refresh` job is wiped!

> We note that this appears to affect _all_ scheduled jobs with the `startClean` flag set; `billing.customer-file-refresh` is simply the only one to be noticed so far.

We were able to consistently replicate this locally by adding a small delay to `water-api` starting up so that it always registers jobs after `background-service`. Conversely, adding a small delay to `background-service` starting up consistently fixed this.

This quick and dirty fix isn't ideal though; yes, we could make `background-service` pause before registering jobs but this is very much tackling the symptom, not the cause, and wouldn't work in a scenario where for some reason `water-api` takes an unexpectedly long time to start up. We therefore look for a more reliable fix.

The solution we have landed on is to amend `QueueManager` so that we only attempt to clear queues if we are running as `background-service`. There are 4 jobs with the `startClean` flag set and these are all scheduled jobs run by `background-service`, so we know this fix is appropriate.