Remove customer-file-refresh job
https://eaflood.atlassian.net/browse/WATER-5174

We had an issue reported that the [Charging Module API](https://github.com/DEFRA/sroc-charging-module-api)

It was noted that after updating the address for a billing account, no transaction file was being generated for SOP by the **Charging Module API**. We resolved that issue, but then spotted that the job to update the transaction reference wasn't running due to a [race condition](https://github.com/DEFRA/water-abstraction-service/pull/2708).

When we investigated, we found that no billing account has been updated with a transaction reference _since October 2024_! The current job in is hardcoded only to look back a day. To update those records missed, we need to go back 11 months.

Therefore, this seemed a prime opportunity to migrate the legacy job to [water-abstraction-system](https://github.com/DEFRA/water-abstraction-system), including making the period the job looks at configurable.

The [new job has been added](https://github.com/DEFRA/water-abstraction-system/pull/2253), so now we can delete the legacy one in this app.