Add missing licenceId to abs alert notifications
https://eaflood.atlassian.net/browse/WATER-5232

Water abstraction alert notifications created by the legacy code contain the relevant `licenceId` in the `water.scheduled_notification` `personalisation` field. Those created by [water-abstraction-system](https://github.com/DEFRA/water-abstraction-system) don't.

This discrepancy has caused the last alert data we show in the 'view licence monitoring stations' page to be incorrect because it was depending on the `licenceId` being present. Turns out it shouldn't have been getting the information in the way it is anyway, but we still want to keep the records consistent.

This one-off data migration adds the missing `licenceId` to those water abstraction alert notification records that don't have it.