Add notification ID to licence gauging stations
https://eaflood.atlassian.net/browse/WATER-5232

We've spotted that the last alert details in the monitoring station page do not align with the ones shown on the view licence monitoring stations page in some cases.

In [Fix last alert info in view lic. mon. station page](https://github.com/DEFRA/water-abstraction-system/pull/2354) we cover what the problem is, and implement a fix.

We also ensured that the water abstraction alert notifications get updated with the [missing licence IDs](https://github.com/DEFRA/water-abstraction-service/pull/2717).

But at the core of the issue is the fact the `water.licence_gauging_stations` table is not linked to the `water.scheduled_notification` table.

The legacy code when creating an abstraction alert would record the type and date in the `licence_gauging_stations` record. When we took over the process, we improved it by recording the details only when we had received confirmation from **Notify** that the email or letter had been sent successfully.

But we still don't actually record the notification ID. The issue we found was in part caused by how complex it was for the new view licence monitoring stations page to find the relevant notification records.

If we captured the notification ID of the last abstraction alert notification sent, these issues would disappear. In fact, `status` and `status_updated` would become redundant (though because that will break our downstream reporting service, removing them will have be done at another time).

This this change, we add a migration to add `notification_id` to the `licence_gauging_stations` table.

We then run a script to populate it for as many of the existing records as possible. Going forward, our change in [water-abstraction-system](https://github.com/DEFRA/water-abstraction-system/pull/2369) will ensure it is populated.