Add notification ID to licence gauging stations (#2719)
https://eaflood.atlassian.net/browse/WATER-5232

We've spotted that the last alert details in the monitoring station page do not align with the ones shown on the view licence monitoring stations page in some cases.

In [Fix last alert info in view lic. mon. station page](https://github.com/DEFRA/water-abstraction-system/pull/2354), we cover what the problem is, and implement a fix.

We also ensured that the water abstraction alert notifications get updated with the [missing licence IDs](https://github.com/DEFRA/water-abstraction-service/pull/2717).

But at the core of the issue is the fact that the `water.licence_gauging_stations` table is not linked to the `water.scheduled_notification` table.

The legacy code, when creating an abstraction alert, would record the type and date in the `licence_gauging_stations` record. When we took over the process, we improved it by recording the details only when we had received confirmation from **Notify** that the email or letter had been sent successfully.

But we still don't actually record the notification ID. The issue we found was in part caused by how complex it was for the new view licence monitoring stations page to find the relevant notification records.

If we captured the notification ID of the last abstraction alert notification sent, these issues would disappear. In fact, `status` and `status_updated` would become redundant (though, because that will break our downstream reporting service, removing them will have to be done at another time).

We update the `licence_gauaging_station` in our **notification-status** job, when it confirms the notification was sent. Its problem is that it has to rely on the notification's `personalisation` JSONB field containing the `licenceGaugingStationId`. Again, we are reliant on a property in a JSONB field rather than a proper column value.

So, this change is primarily about adding columns.

- Add `notification_id` to the `licence_gauging_stations` table
- Add `licence_gauging_station_id` to the `scheduled_notification` table

We've also spotted that we have been consistent in how we populate the JSONB field for new abstraction alerts (another reason to avoid them). We've fixed the functionality in [water-abstraction-system](https://github.com/DEFRA/water-abstraction-system/pull/2369). However, this migration script also corrects the records to restore consistency.

With the fix part done, we add the columns, then populate them for the existing records.