Add `overall_status` and `status_counts` columns to events table (#2728)
https://eaflood.atlassian.net/browse/WATER-5306

The `overall_status` column is populated with the overall status of related notifications. This status is calculated based on the statuses of all notifications associated with an event. It provides a quick reference to the overall status of the event without needing to check each notification individually.

The statuses are calculated as follows:
- If one notification has a status of returned, then show RETURNED
- If no notifications are returned, but one has a status of error, then show ERROR
- If no notifications are returned or errored, but one has a status of pending, then show PENDING
- If no notifications are returned, errored, or pending, then show SENT

The `status_counts` column is a JSONB object that contains counts of each notification status type related to the event.