Fix notification status + amend prev. migration (#2730)
- https://eaflood.atlassian.net/browse/WATER-5306
- https://eaflood.atlassian.net/browse/WATER-5328
- https://eaflood.atlassian.net/browse/WATER-5329

In [Add overall_status and status_counts columns to events table](https://github.com/DEFRA/water-abstraction-service/pull/2728), we added a migration to add two new columns to `water.events`, then populate them based on the existing status of the notifications linked to them.

Then, when working to enable [water-abstraction-system](https://github.com/DEFRA/water-abstraction-system) to view any notification, no matter its status, we came across ones which had a [Notify](https://www.notifications.service.gov.uk/) status of 'cancelled', but a status of 'sent'.

> These are letters where the user has cancelled them in the Notify dashboard before they were sent to the provider.

We realised we needed to capture these in our overall status and counts, but that the existing records would need to be updated before the new migration script kicked in.

However, we then realised we had a bigger issue! Essentially, 'sent' for legacy notifications just means the notification got to Notify. It doesn't mean that Notify has actually sent it. But in the new system, the status of notifications relates to whether Notify has managed to send the notification successfully.

We found thousands of notifications with a status of 'sent', but a Notify status that means error/failure.

This change amends the migration script (that hasn't run in `production` yet!) to include some additional work.

- Update the status of existing notifications to match the 'system' understanding (sent -> error where notify_status is not successful)
- Sort out the date columns. `create` and `modified` now default to the current datetime and don't allow nulls

Then crack on with the previous migration, but with changes.

- Add the new `overall_status` and `status_counts` columns if they don't already exist
- Run the script to determine the overall status and status counts, but now it includes 'cancelled'
- To avoid any issues, update the legacy error count in `water.events` metadata and set the `updated_at`