Fix all paper returns being flagged as ERROR (#2731)
https://eaflood.atlassian.net/browse/WATER-5330

Whilst working on WATER-5329 to [update the statuses for legacy notifications to match how we use them](https://github.com/DEFRA/water-abstraction-service/pull/2730), we spotted that _all_ paper returns created since 2023 have been flagged as `error`.

None of them has a [GOV.UK Notify ID](https://www.notifications.service.gov.uk/) ID or a Notify status, and all have the same error recorded.

```json
{"error":"Notify error","message":"TypeError: Cannot read properties of undefined (reading 'body')"}
```

We tracked it down to a faulty expectation that all responses from Notify have a 'content' property. This is only the case with notifications where we provide the personalisation and use a template.

When a notification is an uploaded PDF, no content is included in the response.

This change fixes the issue.