Flag existing notifications as RETURNED
https://eaflood.atlassian.net/browse/WATER-5334

We've recently [added a callback url for GovNotify](https://github.com/DEFRA/water-abstraction-system/pull/2383) to tell us when a letter is returned to them. We then updated that to [Set set the status and returned at date](https://github.com/DEFRA/water-abstraction-system/pull/2471) on the notification. We can use this information to inform users via the UI which notifications have been returned.

However, this will only tell us about future pings. Notify provide a series of downloadable reports for returned letters which currently go back to June 2023. If we take this information, we can update the status of existing notifications to `RETURNED` as well.

So, this change creates a migration built from the data in those reports. Some notificaations may have been sent via the dashboard directly so won't existing in our service. Others might, but the legacy code errored before the notify ID could be persisted.

Either way, it is better than nothing. The only other wrinkle is when we get a ping, we use the current date as the value for `returned_at`. This isn't available in the reports, so we have used the report date for the value instead.