Add alternative notification id column to scheduled notification (#2737)
https://eaflood.atlassian.net/browse/WATER-5230

As part of the work to adapt our system for dynamic due dates we want to be able to send out a letter in case there are any issues with emailing users. When GovNotify returns an error when we send the returns invitation email we want to then trigger a letter to be sent to the primary user. We want to be able to link the new letter notification to the failed email one.

This PR adds a new column to the scheduled_notification table to link to the new notification.