Use new return log id https://eaflood.atlassian.net/browse/WATER-5353
Work is being carried out to update URLs that use the old return log reference in a query string, e.g. `/system/return-logs?id=v1:9:99/999:10039999:2025-04-01:2026-03-31` to use the new return log UUID as a parameter e.g. `/system/return-logs/3ce89af1-6168-4e92-9673-75cfc87f661e`.

Currently the legacy search code returns a link to the return logs if found which requires updating. In order to update this URL the new return ID needs to be passed to the UI. This change adds the new return ID to the search results so that it can be used by the UI.