Remove the none-working feature flag
Initially, our intent was to stop the check-status job completely.

But on reflection, we may want to show all notifications sent in relation to an account at some point in the future. So, we may as well allow the check-status job to continue checking for the forseeable future.

First step is to remove the redundant feature flag.