Change job timings
We _really_ don't need to be checking the status of notifications every 15 seconds, especially now as none of the data this service handles is visible. So, we now just check twice a day.

We up the send message check from 15 seconds to 30, halving the number of times we'll run the job during the day, yet not making users wait _too_ long else they will think it had failed.