Stop interferring with notices & notifications (#2745)
https://eaflood.atlassian.net/browse/WATER-5395

We knew that the legacy service had a job for updating the status of notifications after they were sent. As we took control of sending notices, this job needed to be disabled.

We thought we'd done that in [Add enable batch notification job feature toggle](https://github.com/DEFRA/water-abstraction-service/pull/2680), but it turns out not to be the case.

We have examples of new fields we've added to the `water.events` table being set to NULL, and fields related to checking the status in `water.scheduled_notification` being populated even though this job is supposed to be disabled.

After checking, it appears that the simple act of registering a 'job', which is configured to repeat, will automatically start it, overriding what we did with `src/lib/queue-manager/start-up-jobs-service.js`.

We need the legacy service to leave all notices (returns invitations and reminders, paper returns and abstraction alerts) alone.

Account-related emails are still sent through the service, so they should be left as is. But as they are not linked to events, the refresh event (`src/modules/batch-notifications/lib/jobs/refresh-event.js`) job is redundant. The check status job (`src/modules/batch-notifications/lib/jobs/check-status.js`) should ignore anything that isn't account-related. The same applies to the send message (`src/modules/batch-notifications/lib/jobs/send-message.js`) job.

So, this change updates the three 'batch-notifications' jobs to only look at account-related notifications.

We also run a migration script to re-populate the `water.events` we think these jobs have messed with.