Soft delete workflow record (#2749)
https://eaflood.atlassian.net/browse/WATER-5407

Our users have pointed our that when they are processing workflow records some of them are being generated again which is causing them to recheck work they've already done. It looks like this is due to the record being hard deleted rather than soft deleted.

This PR changes it to a soft delete so the record is there when the system checks for changes to create workflow records.