Record contact and notice type in notifications (#2751)
https://eaflood.atlassian.net/browse/WATER-5410

This is actually related to our work on dynamic due dates. In the future, due dates will be set against return logs when we get confirmation from Notify that the related notification was successful.

However, we should only be doing this for specific notice and contact types.

What we've realised is that the logic we have for this is faulty, because it only has the `messageRef` to go on. For example, in an ad-hoc notice, you can add a 'single-use' recipient. Because we need to set something for `message_ref`, we've opted to use `returns_invitation_primary_user` for email contacts and `returns_invitation_letter-holder` for letter contacts. These were legacy values used when it handled notices.

But now we've lost the knowledge that this was a single-use contact, not the licence holder or primary user. Plus, we are applying due dates to a return log because it's a successful 'returns invitation'.

Ideally, the two values should be separate; we are sending a **returns invitation** to the **primary user**, or a **paper return** to a **single-use** contact. Then, decisions like which notifications determine whether due dates are applied become much easier to implement.

In this change, we are going to start making this distinct and clear.

- A new field `contact_type` will be added to `water.scheduled_notification`
- A new migration to populate this from existing information
- In the same migration, update `message_ref` to contain only the notice type

As [water-abstraction-system](https://github.com/DEFRA/water-abstraction-system) now handles everything to do with abstraction alert and returns notices, we're ignoring how this change affects legacy functionality (beyond fixing any broken unit tests).