Fix missing communication sent date in external UI
https://eaflood.atlassian.net/browse/WATER-4991

Since we have taken over creating and sending notices, we have been using the `date_created` column instead of the `scheduled_after` column to record when a notification was created.

One of our QA team spotted that when viewing communications in the external UI, that the **Sent date** column was empty for some of them.

We realised that the external UI must still be depending on `scheduled_after`, hence it was missing for recent notifications.

What we hadn't realised until we made this change, was that which notifications it displays is _also_ based on `scheduled_after`! 🤦

This change fixes the query and mapping to ensure 'date' is populated and available to the UI to display in the page.